Add a height index to the ATStates table. DO NOT MERGE.
This is a throwaway branch which can be attempted by those with a high spec CPU, to get an idea on how long this process takes. Only then can we consider adding it as its own case, but I suspect we will need to wait until we have solved the database size issues first.